moved all the game.cpp stuff into interface, it was redundant
